+++ src/routes/settings-assigned-users-route/settings-assigned-users-route.js
@@ -58,7 +58,7 @@ const SettingsAssignedUsersRoute = ({
 
   useEffect(() => {
     return () => clearKBCredentialsUser();
-  }, []);
+  }, [clearKBCredentialsUser]);
 
   const hideAlreadyAssignedMessage = () => {
     setAlreadyAssignedMessageDisplayed(false);