+++ src/routes/settings-route/settings-route.js
@@ -19,6 +19,7 @@ const SettingsRoute = ({
 
   useEffect(() => {
     getKbCredentials();
+  // eslint-disable-next-line react-hooks/exhaustive-deps
   }, []);
 
   const sortKbCredentials = () => {