+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## [7.4.0] (IN PROGRESS)
 * Difficulty returning to the eholdings page after adding a new Agreement. (UIEH-1219)
+* Enable react-hooks/exhaustive-deps ESLint rule. (UIEH-1339)
 
 ## [7.3.2] (https://github.com/folio-org/ui-eholdings/tree/v7.3.2) (2022-11-30)